Add files via upload
Hi Matt,

There are two attachments in this email:

1) project zip file
2) README.pdf file

The project need to use java 8 and maven to build and run. I developed this project in Eclipse IDE and unit test cases are also in the zip. Please let me know if reviewer has any questions.

Thanks
Yue Wu (Henry)